requirements.txt
alabaster==0.7.10
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.5
django-bootstrap==0.2.4
django-localflavor==2.0
docutils==0.14
idna==2.6
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
packaging==17.1
Pillow==5.1.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2018.4
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
